+++ src/Screens/SampleScreen/index.tsx
@@ -44,7 +44,7 @@ export const SampleDataScreens = memo((props: {
       widgetData: removedWidget,
       sync: true,
     }, async () => {
-      await MediaService.deleteMedia({
+      await MediaService.deleteMediaRecursively({
         scope: 'widget',
         id_project: id_project,
         widget: removedWidget,