+++ src/app/index.tsx
@@ -3,10 +3,7 @@ import * as Location from 'expo-location';
 
 import { navigate } from '@Globals/NavigationControler';
 import ConfigService from '@Services/ConfigService';
-import DatabaseService from '@Services/DatabaseService';
-import CredentialService from '@Services/CredentialService';
-import FileSystemService from '@Services/FileSystemService';
-import DocumentFileExportService from '@Services/DocumentFileExportService';
+import { FOLDER_App } from '@Services/FileSystemService';
 
 import { Layout } from '@Layout/index';
 
@@ -28,10 +25,7 @@ export default function Home() {
 }
 
 async function initApp(onFinish: () => void) {
-  await FileSystemService.init();
-  await DocumentFileExportService.createTempFilesFolder();
-  await CredentialService.createCredentialsFolder();
-  await DatabaseService.createDatabaseFolder();
+  await FOLDER_App.init();
   await ConfigService.loadConfig();
   Location.requestForegroundPermissionsAsync();
   onFinish();